+++ lib/sfio/sfexcept.c
@@ -91,7 +91,7 @@ int _sfexcept(Sfio_t * f, int type, ssize_t io, Sfdisc_t * disc)
 
 	/* a normal interrupt, we can continue */
 	errno = 0;
-	f->flags &= ~(SF_EOF | SF_ERROR);
+	f->flags &= (unsigned short)~(SF_EOF | SF_ERROR);
 	SFMTXRETURN(f, SF_ECONT);
     }
